+++ CHANGELOG.md
@@ -5,8 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-
 ## [Unreleased]
+### Added
+- Check for "value" or "path" key in replace patch operations
+- Check for "value" in add patch operations
 
 ## [0.2.2] — 2020-09-18